refactor: Expose AwsV4Signer instead of modules
