+++ README.md
@@ -15,7 +15,7 @@ Most API is simple. But they could be complicated when they are hidden from comp
 ## Quick Start
 
 ```rust
-use reqsign::aws::v4::Signer;
+use reqsign::AwsV4Signer;
 use reqwest::{Client, Request, Url};
 use anyhow::Result;
 
@@ -43,7 +43,7 @@ async fn main() -> Result<()>{
 - Test again official SDK and services.
 - Supported services
   - Aliyun OSS: `reqsign::aliyun::oss::Signer`
-  - AWS services (SigV4): `reqsign::aws::v4::Signer`
+  - AWS services (SigV4): `reqsign::AwsV4Signer`
   - Azure Storage services: `reqsign::azure::storage::Signer`
   - Google services: `reqsign::google::Signer`
   - Huawei Cloud OBS: `reqsign::huaweicloud::obs::Singer`
